[UPDATE] use configuration & put into resource package
