+++ src/main/java/com/immobylette/api/photo/ImmobyletteApiPhotoApplication.java
@@ -2,8 +2,10 @@ package com.immobylette.api.photo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
+@EnableConfigurationProperties
 public class ImmobyletteApiPhotoApplication {
 
     public static void main(String[] args) {
