+++ src/main/resources/application.yml
@@ -17,10 +17,6 @@ spring:
     generate-ddl: false
     show-sql: true
     hibernate.ddl-auto: none
-  gcp:
-    project-id: ${GCP_PROJECT_ID:immobylette}
-    bucket-name: ${GCP_BUCKET_NAME:immobylette--photos}
-    credentials-location: ${GCP_CREDENTIALS_LOCATION:immobylette-credentials.json}
 
 server:
   port: 8080
@@ -28,3 +24,8 @@ server:
     context-path: /api/v1
   error:
     include-message: always
+
+gcs:
+  project-id: ${GCP_PROJECT_ID:immobylette}
+  bucket-name: ${GCP_BUCKET_NAME:immobylette--photos}
+  credentials-location: ${GCP_CREDENTIALS_LOCATION:immobylette-credentials.json}
